Support for invalid frame error handling
When an invalid frame is given, and exception is transported to the higher level. However, if a valid frame is later given, it will be ignored since the class "animating" is still applied to the configurator. It is necessary to catch the exception in update and remove that class, not allowing the recursive action to continue.