feat: call ripe update on configurator instead of calling its own update
For the configurator to update itself, it must call is parent update, which contains checks to prevent two concurrent updates in the configurator. The flushPending call was moved to the end of the changeFrame flow so that it happens only after the end of a change frame. The previous location was causing a deadlock.